requirements.txt
numpy~=1.24.1
opencv-python~=4.7.0.68
matplotlib~=3.6.3
colour-demosaicing~=0.2.2
scikit_image~=0.19.3
tqdm~=4.65.0
pandas~=2.0.0
scikit-learn~=1.2.2
hydra-core~=1.3.2
